+++ src/main/java/insane96mcp/enhancedai/modules/skeleton/feature/SkeletonAIFeature.java
@@ -3,29 +3,37 @@ package insane96mcp.enhancedai.modules.skeleton.feature;
 import insane96mcp.enhancedai.modules.base.ai.AIAvoidEntityGoal;
 import insane96mcp.enhancedai.modules.skeleton.ai.AIRangedBowAttackGoal;
 import insane96mcp.enhancedai.setup.Config;
+import insane96mcp.enhancedai.utils.LogHelper;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.config.BlacklistConfig;
 import insane96mcp.insanelib.utils.IdTagMatcher;
+import insane96mcp.insanelib.utils.RandomHelper;
 import net.minecraft.entity.ai.goal.Goal;
 import net.minecraft.entity.ai.goal.PrioritizedGoal;
 import net.minecraft.entity.monster.AbstractSkeletonEntity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.Tuple;
 import net.minecraftforge.common.ForgeConfigSpec;
+import org.apache.commons.lang3.math.NumberUtils;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Label(name = "Skeleton Shoot", description = "Skeletons are more precise when shooting and strafing is removed, can hit a target from 64 blocks and try to stay away from the target.")
+@Label(name = "Skeleton Shoot", description = "Skeletons are more precise when shooting and strafing is removed, can hit a target from up to 64 blocks and try to stay away from the target.")
 public class SkeletonAIFeature extends Feature {
 
+	private final ForgeConfigSpec.ConfigValue<String> minMaxShootingRangeConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> avoidPlayerChanceConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> arrowInaccuracyConfig;
 	private final BlacklistConfig entityBlacklistConfig;
 
+	private static final String minMaxShootingRangeDefault = "24, 48";
+
+	public Tuple<Integer, Integer> minMaxShootingRange;
 	public double avoidPlayerChance = 0.5d;
 	public double arrowInaccuracy = 2;
 	public ArrayList<IdTagMatcher> entityBlacklist;
@@ -34,6 +42,9 @@ public class SkeletonAIFeature extends Feature {
 	public SkeletonAIFeature(Module module) {
 		super(Config.builder, module);
 		Config.builder.comment(this.getDescription()).push(this.getName());
+		minMaxShootingRangeConfig = Config.builder
+				.comment("The min and max range from where a skeleton will shoot a player.")
+				.define("Min Max Shooting Range", minMaxShootingRangeDefault);
 		avoidPlayerChanceConfig = Config.builder
 				.comment("Chance for a Skeleton to spawn with the ability to avoid the player")
 				.defineInRange("Avoid Player chance", this.avoidPlayerChance, 0d, 1d);
@@ -47,10 +58,30 @@ public class SkeletonAIFeature extends Feature {
 	@Override
 	public void loadConfig() {
 		super.loadConfig();
+		this.minMaxShootingRange = parseIntTuple(this.minMaxShootingRangeConfig.get());
 		this.avoidPlayerChance = this.avoidPlayerChanceConfig.get();
 		this.arrowInaccuracy = this.arrowInaccuracyConfig.get();
 		this.entityBlacklist = IdTagMatcher.parseStringList(this.entityBlacklistConfig.listConfig.get());
 		this.entityBlacklistAsWhitelist = this.entityBlacklistConfig.listAsWhitelistConfig.get();
+	}
+
+	private Tuple<Integer, Integer> parseIntTuple(String s) {
+		String[] split = s.split(",");
+		if (split.length != 2)
+			return new Tuple<>(Integer.parseInt(minMaxShootingRangeDefault.split(",")[0]), Integer.parseInt(minMaxShootingRangeDefault.split(",")[1]));
+		if (!NumberUtils.isParsable(split[0])) {
+			LogHelper.warn("Invalid number for Min Max Shooting Range");
+			return new Tuple<>(Integer.parseInt(minMaxShootingRangeDefault.split(",")[0]), Integer.parseInt(minMaxShootingRangeDefault.split(",")[1]));
+		}
+		int a = Integer.parseInt(split[0]);
+		if (!NumberUtils.isParsable(split[1])) {
+			LogHelper.warn("Invalid number for Min Max Shooting Range");
+			return new Tuple<>(Integer.parseInt(minMaxShootingRangeDefault.split(",")[0]), Integer.parseInt(minMaxShootingRangeDefault.split(",")[1]));
+		}
+		int b = Integer.parseInt(split[1]);
+		if (a > b)
+			a = b;
+		return new Tuple<>(a, b);
 	}
 
 	//TODO Zombies with Ender Pearls
@@ -82,11 +113,11 @@ public class SkeletonAIFeature extends Feature {
 
 		avoidEntityGoals.forEach(skeleton.goalSelector::removeGoal);
 		if (hasAIArrowAttack) {
-			AIRangedBowAttackGoal<AbstractSkeletonEntity> rangedBowAttackGoal = new AIRangedBowAttackGoal<>(skeleton, 1.0d, 20, 64.0f);
+			AIRangedBowAttackGoal<AbstractSkeletonEntity> rangedBowAttackGoal = new AIRangedBowAttackGoal<>(skeleton, 1.0d, 20, RandomHelper.getInt(skeleton.world.rand, minMaxShootingRange.getA(), minMaxShootingRange.getB()));
 			skeleton.goalSelector.addGoal(2, rangedBowAttackGoal);
 
 			if (skeleton.world.rand.nextDouble() < this.avoidPlayerChance) {
-				AIAvoidEntityGoal<PlayerEntity> avoidEntityGoal = new AIAvoidEntityGoal<>(skeleton, PlayerEntity.class, 12.0f, 1.6d, 1.3d);
+				AIAvoidEntityGoal<PlayerEntity> avoidEntityGoal = new AIAvoidEntityGoal<>(skeleton, PlayerEntity.class, 12.0f, 1.5d, 1.25d);
 				skeleton.goalSelector.addGoal(1, avoidEntityGoal);
 			}
 		}